SFC: Add docstrings

+++ SFC/utils.h
@@ -3,17 +3,34 @@
 #include <sstream>
 #include <fstream>
 #include <vector>
+#include <iostream>
 
 #ifndef UTILS_H
 #define UTILS_H
 
+
 namespace utils {
+  /**
+   * Calcualte sigmoid
+   */
   inline double sigmoid(double x) {
     return 1.0 / (1.0 + exp(-x));
   }
+
+  /**
+   * Get a random value in range
+   * @param  around positive range boundary
+   * @return        random number
+   */
   inline double uniform(double around) {
     return -around + (rand() / (RAND_MAX + 1.0) * (2 * around));
   }
+
+  /**
+   * Compute activation based on probability
+   * @param  sample probability
+   * @return        binary activation
+   */
   inline int binomial(double sample) {
     if(sample < 0 || sample > 1) return 0;
 
@@ -21,15 +38,29 @@ namespace utils {
       return 1;
     return 0;
   }
+
+  /**
+   * Pretty print an C-like array of values
+   * @param  data   Array of values
+   * @param  length Size of the array
+   * @return        String representation of the array
+   */
   template<class T> inline std::string print_array(T * data, size_t length) {
     std::string output = "[";
-    for(int x = 0; x < length; x++) output += std::to_string(data[x]) + ", ";
+    for(size_t x = 0; x < length; x++) output += std::to_string(data[x]) + ", ";
     output.erase(output.end()-2, output.end());
     output += "]";
     return output;
   }
+
+  /**
+   * Read array from stringstream
+   * @param ss   Stringstream to read from
+   * @param data Data buffer - array for storing values
+   * @param size Array members to load
+   */
   template<class T> inline void read_array(std::stringstream & ss, T * data, size_t size) {
-    for (int i=0; i<size; i++) {
+    for (size_t i=0; i<size; i++) {
       data[i] = 0;
       if (!ss.good()) continue;
 
@@ -39,6 +70,13 @@ namespace utils {
       value >> data[i] >> std::ws;
     }
   }
+
+
+  /**
+   * Read array from stdin
+   * @param data Data buffer - array for storing values
+   * @param size Array members to load
+   */
   template<class T> inline void read_array(T * data, size_t size) {
     std::string input;
     std::getline(std::cin, input);
@@ -46,6 +84,10 @@ namespace utils {
     read_array(ss, data, size);
   }
 
+  /**
+   * Read a value from stdin
+   * @param data Data storage
+   */
   template<class T> inline void read_value(T & data) {
     std::string input;
     std::getline(std::cin, input);
@@ -53,6 +95,11 @@ namespace utils {
     value >> data;
   }
 
+  /**
+   * Open and parse CSV file
+   * @param filename Name of file to open
+   * @param size     Amount of data on each line
+   */
   inline std::vector<int *> read_csv(std::string filename, size_t size) {
     std::ifstream file(filename);
     std::string line;
@@ -67,6 +114,12 @@ namespace utils {
 
     return samples;
   }
+
+  /**
+   * Check if file exists
+   * @param  filename Name of the file
+   * @return          Boolean value
+   */
   inline bool check_file(std::string filename) {
     std::ifstream infile(filename);
     return infile.good();
